server/src/main/java/com/farmcollector/service/FarmService.java
package com.farmcollector.service;

import com.farmcollector.model.Farm;
import com.farmcollector.repository.FarmRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.List;
import java.util.Optional;

@Service
public class FarmService {

    @Autowired
    private FarmRepository farmRepository;

    public List<Farm> getAllFarms() {
        return farmRepository.findAll();
    }

    public Optional<Farm> getFarmById(Long farmId) {
        return farmRepository.findById(farmId);
    }

    public Farm saveFarm(Farm farm) {
        return farmRepository.save(farm);
    }

    public void deleteFarm(Long farmId) {
        farmRepository.deleteById(farmId);
    }
}
